+++ .vscode/launch.json
@@ -171,7 +171,7 @@
             "type": "node",
             "request": "launch",
             "name": "Start API",
-            "program": "${workspaceFolder}/built/api/couch-proxy.js",
+            "program": "${workspaceFolder}/built/api/api_server.js",
             "preLaunchTask": "npm: build",
         },
         {
