+++ types/api/orders.d.ts
@@ -1,0 +1,19 @@
+
+interface Order {
+  "ownerAddress": string,
+  "escrowAddress": string,
+  "algoAmount": number,
+  "asaAmount": number,
+  "assetLimitPriceD": number,
+  "assetLimitPriceN": number,
+  "asaPrice": number,
+  "formattedPrice": number,
+  "formattedASAAmount": number,
+  "round": number,
+  "unix_time": number,
+  "decimals": number,
+  "version": string,
+  "isAlgoBuyEscrow": boolean,
+  "assetId": number
+}
+
